Create deploy-canister-in-rust
Here is the complete script to deploy an Interledger Connector Protocol (ICP) canister in Rust.

This script defines a deploy_connector function that uses the ICP Connectors API to deploy a new canister. The function takes as input an instance of the Connectors API, the canister's ILP address, and the name of the backend to use. The function sends a POST request to the API to deploy the canister, and then displays a success or error message based on the API's response.

The main function uses the deploy_connector function to deploy a canister with the ILP address "example.connector" and the backend "example_backend". Be sure to replace these values with ones that are suitable for your development environment.

It's important to note that this script is a simple example and can be modified and customized to fit your needs. You can, for example, add additional configuration options, handle errors in more detail, etc. For more information on the ICP Connectors API and other features